#308 Do not apply position as default sort field
When the flat catalog is disabled, sorting the product collection by
position had no effect on the SQL query.
When the flat catalog is enabled, sorting the product collection by
position causes the issue explained in issue #308.

Because it didn't have any effect when the flat catalog is disabled,
removing position as the default sort order field for SSR product
sliders is a backward compatible change.

Closes #308